project dir restructure
