+++ inference_perf/client/__init__.py
@@ -1,4 +1,4 @@
-# Copyright 2025
+# Copyright 2025 The Kubernetes Authors.
 #
 # Licensed under the Apache License, Version 2.0 (the "License");
 # you may not use this file except in compliance with the License.
@@ -11,32 +11,8 @@
 # WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
 # See the License for the specific language governing permissions and
 # limitations under the License.
-from abc import ABC, abstractmethod
-from dataset import InferenceData
-from reportgen import ReportGenerator, Metric
+from .base import ModelServerClient
+from .mock_client import MockModelServerClient
 
 
-class Client(ABC):
-    @abstractmethod
-    def __init__(self, *args) -> None:
-        pass
-
-    @abstractmethod
-    def set_report_generator(self, reportgen: ReportGenerator) -> None:
-        self.reportgen = reportgen
-
-    @abstractmethod
-    def process_request(self, data: InferenceData) -> None:
-        raise NotImplementedError
-
-
-class MockModelServerClient(Client):
-    def __init__(self, uri: str) -> None:
-        self.uri = uri
-
-    def set_report_generator(self, reportgen: ReportGenerator) -> None:
-        self.reportgen = reportgen
-
-    def process_request(self, data: InferenceData) -> None:
-        print("Processing request - " + data.system_prompt)
-        self.reportgen.collect_metrics(Metric(name=data.system_prompt))
+__all__ = ["ModelServerClient", "MockModelServerClient"]
