+++ inference_perf/client/base.py
@@ -1,4 +1,4 @@
-# Copyright 2025
+# Copyright 2025 The Kubernetes Authors.
 #
 # Licensed under the Apache License, Version 2.0 (the "License");
 # you may not use this file except in compliance with the License.
@@ -11,3 +11,21 @@
 # WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
 # See the License for the specific language governing permissions and
 # limitations under the License.
+from abc import ABC, abstractmethod
+from datagen import InferenceData
+from reportgen import ReportGenerator
+from typing import Tuple
+
+
+class ModelServerClient(ABC):
+    @abstractmethod
+    def __init__(self, *args: Tuple[int, ...]) -> None:
+        pass
+
+    @abstractmethod
+    def set_report_generator(self, reportgen: ReportGenerator) -> None:
+        self.reportgen = reportgen
+
+    @abstractmethod
+    def process_request(self, data: InferenceData) -> None:
+        raise NotImplementedError
