+++ inference_perf/metrics/client.py
@@ -3,5 +3,5 @@ from abc import ABC, abstractmethod
 
 class Metrics_Client(ABC):
     @abstractmethod
-    def __init__(self):
+    def __init__(self) -> None:
         pass
