+++ inference_perf/reportgen/mock_reportgen.py
@@ -11,3 +11,17 @@
 # WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
 # See the License for the specific language governing permissions and
 # limitations under the License.
+from .base import ReportGenerator, Metric
+from typing import List
+
+
+class MockReportGenerator(ReportGenerator):
+    def __init__(self) -> None:
+        self.metrics: List[Metric] = []
+
+    def collect_metrics(self, metric: Metric) -> None:
+        self.metrics.append(metric)
+
+    def generate_report(self) -> None:
+        print("\n\nGenerating Report ..")
+        print("Report: Total Requests = " + str(len(self.metrics)))
